Move ansible auth methods to a mixin and constantize the auth types
